Moved code from __main__.py into main.py so it can be imported elsewhere.
